requirements.txt
mcp
httpx
python-dotenv
langgraph
langchain-community
dashscope
langchain-mcp-adapters
langchain-core
fastapi
uvicorn
